move Neural data classes to separate module
